+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Get-EC2EndpointConnection-Cmdlet.cs
@@ -44,7 +44,7 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         #region Parameter Filter
         /// <summary>
         /// <para>
-        /// <para>One or more filters.</para><ul><li><para><code>service-id</code> - The ID of the service.</para></li><li><para><code>vpc-endpoint-owner</code> - The ID of the Amazon Web Services account ID that
+        /// <para>One or more filters.</para><ul><li><para><code>ip-address-type</code> - The IP address type (<code>ipv4</code> | <code>ipv6</code>).</para></li><li><para><code>service-id</code> - The ID of the service.</para></li><li><para><code>vpc-endpoint-owner</code> - The ID of the Amazon Web Services account ID that
         /// owns the endpoint.</para></li><li><para><code>vpc-endpoint-state</code> - The state of the endpoint (<code>pendingAcceptance</code>
         /// | <code>pending</code> | <code>available</code> | <code>deleting</code> | <code>deleted</code>
         /// | <code>rejected</code> | <code>failed</code>).</para></li><li><para><code>vpc-endpoint-id</code> - The ID of the endpoint.</para></li></ul>