+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Get-EC2VpcEndpoint-Cmdlet.cs
@@ -43,7 +43,7 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         #region Parameter Filter
         /// <summary>
         /// <para>
-        /// <para>One or more filters.</para><ul><li><para><code>service-name</code> - The name of the service.</para></li><li><para><code>vpc-id</code> - The ID of the VPC in which the endpoint resides.</para></li><li><para><code>vpc-endpoint-id</code> - The ID of the endpoint.</para></li><li><para><code>vpc-endpoint-state</code> - The state of the endpoint (<code>pendingAcceptance</code>
+        /// <para>One or more filters.</para><ul><li><para><code>ip-address-type</code> - The IP address type (<code>ipv4</code> | <code>ipv6</code>).</para></li><li><para><code>service-name</code> - The name of the service.</para></li><li><para><code>vpc-id</code> - The ID of the VPC in which the endpoint resides.</para></li><li><para><code>vpc-endpoint-id</code> - The ID of the endpoint.</para></li><li><para><code>vpc-endpoint-state</code> - The state of the endpoint (<code>pendingAcceptance</code>
         /// | <code>pending</code> | <code>available</code> | <code>deleting</code> | <code>deleted</code>
         /// | <code>rejected</code> | <code>failed</code>).</para></li><li><para><code>vpc-endpoint-type</code> - The type of VPC endpoint (<code>Interface</code>
         /// | <code>Gateway</code> | <code>GatewayLoadBalancer</code>).</para></li><li><para><code>tag</code>:&lt;key&gt; - The key/value combination of a tag assigned to the