+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/New-EC2VpcEndpoint-Cmdlet.cs
@@ -43,6 +43,28 @@ namespace Amazon.PowerShell.Cmdlets.EC2
     public partial class NewEC2VpcEndpointCmdlet : AmazonEC2ClientCmdlet, IExecutor
     {
         
+        #region Parameter DnsOptions_DnsRecordIpType
+        /// <summary>
+        /// <para>
+        /// <para>The DNS records created for the endpoint.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [AWSConstantClassSource("Amazon.EC2.DnsRecordIpType")]
+        public Amazon.EC2.DnsRecordIpType DnsOptions_DnsRecordIpType { get; set; }
+        #endregion
+        
+        #region Parameter IpAddressType
+        /// <summary>
+        /// <para>
+        /// <para>The IP address type for the endpoint.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        [AWSConstantClassSource("Amazon.EC2.IpAddressType")]
+        public Amazon.EC2.IpAddressType IpAddressType { get; set; }
+        #endregion
+        
         #region Parameter PolicyDocument
         /// <summary>
         /// <para>
@@ -240,6 +262,8 @@ namespace Amazon.PowerShell.Cmdlets.EC2
             }
             #pragma warning restore CS0618, CS0612 //A class member was marked with the Obsolete attribute
             context.ClientToken = this.ClientToken;
+            context.DnsOptions_DnsRecordIpType = this.DnsOptions_DnsRecordIpType;
+            context.IpAddressType = this.IpAddressType;
             context.PolicyDocument = this.PolicyDocument;
             context.PrivateDnsEnabled = this.PrivateDnsEnabled;
             if (this.RouteTableId != null)
@@ -292,6 +316,29 @@ namespace Amazon.PowerShell.Cmdlets.EC2
             if (cmdletContext.ClientToken != null)
             {
                 request.ClientToken = cmdletContext.ClientToken;
+            }
+            
+             // populate DnsOptions
+            var requestDnsOptionsIsNull = true;
+            request.DnsOptions = new Amazon.EC2.Model.DnsOptionsSpecification();
+            Amazon.EC2.DnsRecordIpType requestDnsOptions_dnsOptions_DnsRecordIpType = null;
+            if (cmdletContext.DnsOptions_DnsRecordIpType != null)
+            {
+                requestDnsOptions_dnsOptions_DnsRecordIpType = cmdletContext.DnsOptions_DnsRecordIpType;
+            }
+            if (requestDnsOptions_dnsOptions_DnsRecordIpType != null)
+            {
+                request.DnsOptions.DnsRecordIpType = requestDnsOptions_dnsOptions_DnsRecordIpType;
+                requestDnsOptionsIsNull = false;
+            }
+             // determine if request.DnsOptions should be set to null
+            if (requestDnsOptionsIsNull)
+            {
+                request.DnsOptions = null;
+            }
+            if (cmdletContext.IpAddressType != null)
+            {
+                request.IpAddressType = cmdletContext.IpAddressType;
             }
             if (cmdletContext.PolicyDocument != null)
             {
@@ -391,6 +438,8 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         internal partial class CmdletContext : ExecutorContext
         {
             public System.String ClientToken { get; set; }
+            public Amazon.EC2.DnsRecordIpType DnsOptions_DnsRecordIpType { get; set; }
+            public Amazon.EC2.IpAddressType IpAddressType { get; set; }
             public System.String PolicyDocument { get; set; }
             public System.Boolean? PrivateDnsEnabled { get; set; }
             public List<System.String> RouteTableId { get; set; }