+++ modules/AWSPowerShell/Cmdlets/SecretsManager/Basic/Get-SECSecretList-Cmdlet.cs
@@ -39,8 +39,8 @@ namespace Amazon.PowerShell.Cmdlets.SEC
     /// To get the secret value from <code>SecretString</code> or <code>SecretBinary</code>,
     /// call <a>GetSecretValue</a>.
     /// </para><para>
-    /// For information about finding secrets in the console, see <a href="https://docs.aws.amazon.com/secretsmanager/latest/userguide/manage_search-secret.html">Enhanced
-    /// search capabilities for secrets in Secrets Manager</a>.
+    /// For information about finding secrets in the console, see <a href="https://docs.aws.amazon.com/secretsmanager/latest/userguide/manage_search-secret.html">Find
+    /// secrets in Secrets Manager</a>.
     /// </para><para><b>Required permissions: </b><code>secretsmanager:ListSecrets</code>. For more information,
     /// see <a href="https://docs.aws.amazon.com/secretsmanager/latest/userguide/reference_iam-permissions.html#reference_iam-permissions_actions">
     /// IAM policy actions for Secrets Manager</a> and <a href="https://docs.aws.amazon.com/secretsmanager/latest/userguide/auth-and-access.html">Authentication