+++ modules/AWSPowerShell/Cmdlets/SecretsManager/Basic/Get-SECSecretVersionIdList-Cmdlet.cs
@@ -28,14 +28,13 @@ using Amazon.SecretsManager.Model;
 namespace Amazon.PowerShell.Cmdlets.SEC
 {
     /// <summary>
-    /// Lists the versions for a secret. 
+    /// Lists the versions of a secret. Secrets Manager uses staging labels to indicate the
+    /// different versions of a secret. For more information, see <a href="https://docs.aws.amazon.com/secretsmanager/latest/userguide/getting-started.html#term_version">
+    /// Secrets Manager concepts: Versions</a>.
     /// 
     ///  
     /// <para>
     /// To list the secrets in the account, use <a>ListSecrets</a>.
-    /// </para><para>
-    /// To get the secret value from <code>SecretString</code> or <code>SecretBinary</code>,
-    /// call <a>GetSecretValue</a>.
     /// </para><para><b>Required permissions: </b><code>secretsmanager:ListSecretVersionIds</code>. For
     /// more information, see <a href="https://docs.aws.amazon.com/secretsmanager/latest/userguide/reference_iam-permissions.html#reference_iam-permissions_actions">
     /// IAM policy actions for Secrets Manager</a> and <a href="https://docs.aws.amazon.com/secretsmanager/latest/userguide/auth-and-access.html">Authentication